+++ src/app/locales/en/components/content-clusters-spaces.js
@@ -1,0 +1,8 @@
+export default {
+  configurationChanged: 'Configuration of "{{{spaceName}}}" space support has ' +
+    'been changed',
+  configurationAction: 'configuration of {{{spaceName}}}',
+  revocationAction: 'space support revocation',
+  revokeSuccess: 'Support for space "{{{spaceName}}}" has been revoked',
+  supportSuccess: 'Added a new support for space',
+};