+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-3870 Show notify after space support settings change
 * VFS-3706 Do not allow to enter improper routes in panel (fixes also VFS-3895: PATCH request after provider deregistration)
 * VFS-3928 Less restrictive validation of provider/onezone domain name
 * VFS-3592 Added common favicon